Refactored pageNumber
The changes are the same for Public timeline and User timeline.

Initially we checked the pageNumber for several conditions inside the .cshtml file. Now it gets handled inside the OnGet function in the .cshtml.cs file.

I have created a get/set variable for the pageNumber inside the .cshtml.cs file. This variable can be accessed inside the .cshtml file. This means we only have to handle errors (like wrong datatype, valid page etc) inside the .cshtml.cs file.