Working follower lists
Reused already existing method in AuthorRepository GetFollowedAuthors(), but i needed another method to get the authors that followed a specific user.

For that i could have used the same thought process as before, and added a list to our core Author.cs called FollowedBy or something, but this would mean new migrations and changes to the DB schema. I found this to be a bad idea as other branches are currently making changes to the schema and i therefore did not want to risk conflicts. I therefore created a method where i iterate through our list of authors O(n), and used contains O(n) which for this project seemed acceptable as our user count is low.