Bugfix for constraint bug in UserTimeline
There was an bug in the UserTimeline when a user tried to send a cheep with no text. Instead of coming up with the error message, it would instead encounter a NullReferenceException. This was because the PageAuthor, FollowingList and FollowingMeList variables were not set in the OnPost method. Which was an issue when a constraint was met